requirements.txt
fastapi
uvicorn
polygon-api-client
pandas
numpy
scikit-learn
tensorflow
